edge/requirements.txt
requests
python-dateutil
uvicorn[standard]==0.15.0
gunicorn==20.1.0
fastapi==0.68.1